feat: rename eventQueue to eventEmitter, improve API
